modified user

+++ ecarpo-bms-imp-server/ecarpo-bms-imp-server-api/src/main/java/com/ecarpo/bms/eas/server/user/service/EasUserService.java
@@ -16,6 +16,6 @@ public interface EasUserService {
   
   ResultDTO<Integer> login(UserPwdDTO dto);
   
-  ResultDTO<IdNameQO> getStoresByUserId(Long userId);
+  ResultDTO<IdNameQO> getStoresByUserId(Integer userId);
   
 }
