+++ ecarpo-bms-imp-server/ecarpo-bms-imp-server-provider/src/main/java/com/ecarpo/bms/eas/server/user/service/IUserManager.java
@@ -40,7 +40,7 @@ public class IUserManager extends BaseManager<IUserMapper, IUserDO> {
     return (l != null && l == 1) ? true : false;
   }
   
-  public IdNameQO getStoresByUserId(Long userId) {
+  public IdNameQO getStoresByUserId(Integer userId) {
     return mapper.getStoresByUserId(userId);
   }
